Improved: UI for the test drive page, and moved the duplicate brokering and reset logic in a single component(#254)
